Fix issue with IHX generation
Switching from SDCC to stm8-objcopy to Intel HEX file generation
has the advantage of just copying the relevant sections (and ignoring
the debug sections that sdcc would otherwise include), but also
care must be taken that all sections which do not end up in FLASH
are removed.

Two more sections are removed during copying, which are created
when initialized and uninitialized global variables are used
within a project.