Add simple UART printf example, refactor old one
The old "spl-uart" example was a rather complex UART1 to UART3
loopback and buffer check firmware, that also doesn't compile for the
boards that do not have a UART1 peripheral.

This renames the old exapmle to spl-uart-loopback and adds a new,
simple UART printf() example which works for all supported STM8 boards.